moved interfaces to separate files, added constructors, renaming grids
