+++ Alligator.DataLayer/Repositories/OrderDetailRepository.cs
@@ -10,16 +10,6 @@ using System.Threading.Tasks;
 
 namespace Alligator.DataLayer.Repositories
 {
-    public interface IOrderDetailRepository
-    {
-        void AddOrderDetail(int amount, int orderId, int productId);
-        void DeleteOrderDetail(int id);
-        void DeleteOrderDetailByProductId(int id);
-        void EditOrderDetail(int id, int amount);
-        OrderDetail GetOrderDetailById(int id);
-        List<OrderDetail> GetOrderDetailsByOrderId(int id);
-        List<OrderDetail> GetOrderDetailsByProductId(int id);
-    }
 
     public class OrderDetailRepository : BaseRepository, IOrderDetailRepository
     {
