+++ Alligator.DataLayer/Repositories/OrderRepository.cs
@@ -10,16 +10,6 @@ using System.Threading.Tasks;
 
 namespace Alligator.DataLayer.Repositories
 {
-    public interface IOrderRepository
-    {
-        void AddOrder(DateTime date, int clientId, string address);
-        void DeleteOrder(int id);
-        void EditOrder(DateTime date, int id, string address);
-        List<Order> GetAllOrders();
-        Order GetOrderById(int id);
-        List<Order> GetOrdersByClientId(int id);
-    }
-
     public class OrderRepository : BaseRepository, IOrderRepository
     {
