+++ Alligator/TabItems/TabItemOrders.xaml
@@ -12,7 +12,7 @@
 
     <Grid>
         <!--Первое окно Все заказы-->
-        <Grid Visibility="{Binding AllOrdersWindow}" Background="#C4C4C4">
+        <Grid Visibility="{Binding OrdersWindowVisibility}" Background="#C4C4C4">
             <Grid.ColumnDefinitions>
             </Grid.ColumnDefinitions>
         <Grid.RowDefinitions>
@@ -45,7 +45,7 @@
         </Grid>
 
 <!--Второе окно Детали заказа-->
-        <Grid VerticalAlignment="Stretch" Height="Auto" Margin="5,5,5,5" Visibility="{Binding OrdersInfoWindow}">
+        <Grid VerticalAlignment="Stretch" Height="Auto" Margin="5,5,5,5" Visibility="{Binding OrdersInfoWindowVisibility}">
                 <Grid VerticalAlignment="Stretch" Height="Auto" Margin="5,5,5,5">
                     <Grid.ColumnDefinitions>
                         <ColumnDefinition Width="1*"/>
@@ -141,7 +141,7 @@
             </Grid>
 
             <!--третье окно Добавить заказ-->
-        <Grid Visibility="{Binding AllOrdersWindow}">
+        <Grid Visibility="{Binding AddOrderWindowVisibility}">
             <Grid VerticalAlignment="Stretch" Height="Auto" Margin="5,5,5,5">
                 <Grid VerticalAlignment="Stretch" Height="Auto" Margin="5,5,5,5">
                     <Grid.ColumnDefinitions>
