+++ Alligator/TabItems/TabItemOrders.xaml.cs
@@ -17,8 +17,8 @@ namespace Alligator.UI.TabItems
             InitializeComponent();
             _viewModel = new TabItemOrdersViewModel();
             DataContext = _viewModel;
-            _viewModel.AddOrderWindow = Visibility.Collapsed;
-            _viewModel.OrdersInfoWindow = Visibility.Collapsed;
+            _viewModel.AddOrderWindowVisibility = Visibility.Collapsed;
+            _viewModel.OrdersInfoWindowVisibility = Visibility.Collapsed;
         }
     }
 }
